Update socs/agents/wiregrid_actuator/agent.py
remove unnecessary casting to float.

+++ socs/agents/wiregrid_actuator/agent.py
@@ -246,7 +246,7 @@ class WiregridActuatorAgent:
         # No limit switch check during this moving
         # Only this function reflects pos/neg of <distance>.
         # Please use +/- distance for insertion/ejection, respectively.
-        ret = self.actuator.move(0.5 * (float)(1. if is_insert else -1.), speedrate=0.2)
+        ret = self.actuator.move(0.5 * (1. if is_insert else -1.), speedrate=0.2)
         if not ret:
             msg = '_insert_eject(): ERROR!:'\
                   'Failed to move in the additional small moving after the last slow moving.'
